Minor fixes

+++ database/filter.db.php
@@ -1,6 +1,8 @@
 <?php
 
 declare(strict_types=1);
+require_once(__DIR__ . '/../utils/validation.php');
+
 
 function getEntitiesFromType(PDO $db, $type) {
     $stmt = $db->prepare("SELECT * FROM $type");
@@ -9,11 +11,16 @@ function getEntitiesFromType(PDO $db, $type) {
 }
 
 function removeEntity(PDO $db,$entity,$type) {
+    if(!is_valid_entity($type)) {
+        exit();
+    }
     $stmt = $db->prepare("DELETE FROM $type WHERE id=?");
     $stmt->execute([$entity]);
 }
-
 function addEntity(PDO $db, $entity,$type) {
+    if(!is_valid_entity($type)) {
+        exit();
+    }
     $stmt = $db->prepare("INSERT INTO $type (name) SELECT (?) WHERE NOT EXISTS (SELECT name FROM $type WHERE name = ?)");
     $stmt->execute([$entity,$entity]);
 }
